Add files via upload
Run BuildTables to build the 2 tables in the "partnership" MYSQL database (must already have created). Will need to adjust user and password fields to for connector to work with your MYSQL database.

Run Partnership to have all the users' data imported to the User table by calling Extraction (only calls Extraction first time Partnership is run/ if User table is empty). Partnership then runs our system starting with user login. If user doesn't have a partner ("unselected") we go to a menu asking if they want to send and invitation, view (and possibly accept) an invitation, or logout. If user is in a partnership ("selected") we go to a screen greeting them and telling them who their partner is (non-distinct name), and gives them an option to logout.

Extraction is called the first time Partnership is run to extract the user's data into the MYSQL User table

Trans class is used to connect with our MYSQL database using JDBC connector.